requirements.txt
mysql-connector-python
neo4j
colorama